pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ddns-traefik-sync"
version = "0.1.0"
description = "Sync Cloudflare A records to domains discovered from Traefik dynamic config files."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0.1",
  "requests>=2.32.3",
]

[project.scripts]
ddns-traefik-sync = "app.main:main"

[tool.setuptools]
packages = ["app"]
